Name the setting 'legacy_runner'
Basically, the old runner is now called LegacyDeferringTextTestRunner,
and the former FastDeferringTextTestRunner is the new DeferrableTestCase.
