+++ unittesting.json
@@ -3,7 +3,7 @@
     "pattern" : "test*.py",
     "async": false,
     "deferred": true,
-    "fast_timings": true,
+    "legacy_runner": false,
     "verbosity": 2,
     "capture_console": false,
     "reload_package_on_testing": true,
