+++ unittesting/core/__init__.py
@@ -1,13 +1,13 @@
 from .st3.runner import DeferringTextTestRunner, AWAIT_WORKER
-from .st3.fast_runner import FastDeferringTextTestRunner
+from .st3.legacy_runner import LegacyDeferringTextTestRunner
 from .st3.case import DeferrableTestCase
 from .st3.suite import DeferrableTestSuite
 from .loader import UnitTestingLoader as TestLoader
 
 __all__ = [
     "TestLoader",
-    "FastDeferringTextTestRunner",
     "DeferringTextTestRunner",
+    "LegacyDeferringTextTestRunner",
     "DeferrableTestCase",
     "DeferrableTestSuite",
     "AWAIT_WORKER"
